Refactoring the notion of documents and collections.
