+++ src/main/java/io/anserini/collection/CW12Collection.java
@@ -1,4 +1,4 @@
-package io.anserini.document;
+package io.anserini.collection;
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one or more
@@ -17,11 +17,28 @@ package io.anserini.document;
  * limitations under the License.
  */
 
-/**
- * Indexable Interface
- */
-public interface Indexable {
-  String id();
-  String content();
-  boolean indexable();
+import io.anserini.document.ClueWeb12WarcRecord;
+import io.anserini.document.SourceDocument;
+
+import java.io.IOException;
+
+public class CW12Collection<D extends ClueWeb12WarcRecord> extends WarcCollection {
+  public CW12Collection() throws IOException {
+    super();
+  }
+
+  @Override
+  public SourceDocument next() {
+    ClueWeb12WarcRecord doc = new ClueWeb12WarcRecord();
+    try {
+      doc = (D)doc.readNextWarcRecord(inStream, ClueWeb12WarcRecord.WARC_VERSION);
+      if (doc == null) {
+        at_eof = true;
+        doc = null;
+      }
+    } catch (IOException e1) {
+      doc = null;
+    }
+    return doc;
+  }
 }
