+++ src/main/java/io/anserini/collection/DiscoverFiles.java
@@ -1,4 +1,4 @@
-package io.anserini.index.collections;
+package io.anserini.collection;
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one or more
