+++ src/main/java/io/anserini/collection/Gov2Collection.java
@@ -1,14 +1,11 @@
-package io.anserini.index.collections;
-
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
+/**
+ * Anserini: An information retrieval toolkit built on Lucene
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,8 +14,10 @@ package io.anserini.index.collections;
  * limitations under the License.
  */
 
-import io.anserini.document.Gov2Record;
-import io.anserini.document.Indexable;
+package io.anserini.collection;
+
+import io.anserini.document.Gov2Document;
+import io.anserini.document.SourceDocument;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -32,7 +31,7 @@ import java.util.Arrays;
 import java.util.HashSet;
 import java.util.zip.GZIPInputStream;
 
-public class Gov2Collection<D extends Gov2Record> extends TrecCollection {
+public class Gov2Collection<D extends Gov2Document> extends TrecCollection {
   public Gov2Collection() throws IOException {
     super();
     skippedFilePrefix = new HashSet<>();
@@ -49,8 +48,8 @@ public class Gov2Collection<D extends Gov2Record> extends TrecCollection {
   }
 
   @Override
-  public Indexable next() {
-    Gov2Record doc = new Gov2Record();
+  public SourceDocument next() {
+    Gov2Document doc = new Gov2Document();
     try {
       doc = (D) doc.readNextRecord(bRdr);
       if (doc == null) {
