+++ src/main/java/io/anserini/document/TrecDocument.java
@@ -1,14 +1,11 @@
-package io.anserini.document;
-
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
+/**
+ * Anserini: An information retrieval toolkit built on Lucene
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,10 +14,15 @@ package io.anserini.document;
  * limitations under the License.
  */
 
+package io.anserini.document;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 
-public class TrecRecord implements Indexable {
+/**
+ * A TREC document.
+ */
+public class TrecDocument implements SourceDocument {
 
   protected final String DOCNO = "<DOCNO>";
   protected final String TERMINATING_DOCNO = "</DOCNO>";
@@ -40,7 +42,7 @@ public class TrecRecord implements Indexable {
   protected String id;
   protected String content;
 
-  public Indexable readNextRecord(BufferedReader reader) throws IOException {
+  public SourceDocument readNextRecord(BufferedReader reader) throws IOException {
     StringBuilder builder = new StringBuilder();
     boolean found = false;
     int inTag = -1;
@@ -94,7 +96,7 @@ public class TrecRecord implements Indexable {
     }
   }
 
-  public Indexable parseRecord(StringBuilder builder) {
+  public SourceDocument parseRecord(StringBuilder builder) {
     int i = builder.indexOf(DOCNO);
     if (i == -1) throw new RuntimeException("cannot find start tag " + DOCNO);
