--- src/main/java/io/anserini/index/collections/CollectionClass.java
@@ -1,8 +1,0 @@
-package io.anserini.index.collections;
-
-/**
- * Information Retrieval Collections
- */
-public enum CollectionClass {
-  CW09, CW12, GOV2, Trec
-}
